refactor(ui): commons extraction
