+++ aether-ui/aether/ui/assets/apps/components/index.jsx
@@ -1,9 +1,11 @@
 // import all components
 import AppLayout from './AppLayout'
 import NavBar from './NavBar'
+import * as Common from './common'
 
 // export all components
 export {
   AppLayout,
-  NavBar
+  NavBar,
+  Common
 }
